refactor(llm): share x-request-id context helper
